+++ Inc/accelero.h
@@ -22,7 +22,7 @@ typedef struct
     void      (*ClearIT)(void);
     void      (*FilterConfig)(uint8_t);
     void      (*FilterCmd)(uint8_t);
-    void      (*GetXYZ)(int16_t *);
+    void      (*GetXYZ)(float *);
 } ACCELERO_DrvTypeDef;
 
 /**
